--- a/optuna/study/study.py
+++ b/optuna/study/study.py
@@ -1,4 +1,5 @@
 import copy
+from datetime import datetime, timedelta
 from numbers import Real
 import threading
 from typing import Any
@@ -54,14 +55,18 @@ _logger = logging.get_logger(__name__)
 
 
 class AllTrialsCache:
+    CACHE_TTL = timedelta(minutes=1)
+
     _is_active: bool
     _cache: Optional[List["FrozenTrial"]]
     _getter: Optional[Callable[[], List["FrozenTrial"]]]
+    _last_updated: datetime
 
     def __init__(self) -> None:
         self._is_active = False
         self._cache = None
         self._getter = None
+        self._last_updated = datetime.now()
 
     def is_active(self) -> bool:
         return self._is_active
@@ -77,9 +82,10 @@ class AllTrialsCache:
 
     def get(self) -> List["FrozenTrial"]:
         assert self._is_active
-        if self._cache is None:
+        if self._cache is None or (datetime.now() - self._last_updated > AllTrialsCache.CACHE_TTL):
             assert self._getter is not None
             self._cache = self._getter()
+            self._last_updated = datetime.now()
         return self._cache