@amylase 2023-03-02
This issue will be resolved in https://github.com/optuna/optuna/pull/4468 with the different approach.