@amylase 2023-03-02
Measured the performance of this change under the same settings as [this experiment](https://gist.github.com/c-bata/16d654471393f59e064482ffb8b52bf8) in my computer.

| branch / storage | RDBStorage with CachedStorage | RDBStorage |
|-|-|-|
| master | 68.78 sec | 194.87 sec |
|-|-|-|
| this PR | 50.12 sec (1.37x faster) | 70.96 sec (2.75x faster) |